+++ src/core/__tests__/config.module.spec.ts
@@ -1,5 +1,6 @@
 import { ConfigModule, ConfigService } from '@nestjs/config'
 import { Test } from '@nestjs/testing'
+import { beforeEach, describe, expect, it } from 'vitest'
 
 import { configModule } from '~/core/config.module'