+++ vitest.config.ts
@@ -1,0 +1,22 @@
+import swc from 'unplugin-swc'
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+  test: {
+    globals: true,
+    coverage: {
+      provider: 'v8',
+      reportsDirectory: './coverage',
+    },
+  },
+  plugins: [
+    swc.vite({
+      module: { type: 'es6' },
+    }),
+  ],
+  resolve: {
+    alias: {
+      '~': '/src',
+    },
+  },
+})